Fixed bugs in builtin functions
Change the return type from int to int64_t

+++ builtin/builtin.c
@@ -3,9 +3,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-void* __alloc(size_t sz) {
-  return malloc(sz);
-}
+void *__alloc(size_t sz) { return malloc(sz); }
 
 void *__string__add(void *lhs, void *rhs) {
   uint64_t lhs_len = *((int64_t *)lhs - 1);
@@ -22,7 +20,7 @@ void *__string__add(void *lhs, void *rhs) {
 }
 
 void *__string__substring(void *ptr, int64_t left, int64_t right) {
-  int len = right - left + 1;
+  int64_t len = right - left + 1;
   char *res_inst_ptr = (char *)malloc(len + 8);
   *(int64_t *)(res_inst_ptr) = len;
   res_inst_ptr += 8;
@@ -49,7 +47,7 @@ int64_t __string__equal(void *lhs, void *rhs) {
   uint64_t rhs_len = *((int64_t *)rhs - 1);
   if (lhs_len != rhs_len)
     return 0;
-  int res = memcmp(lhs, rhs, lhs_len);
+  int64_t res = memcmp(lhs, rhs, lhs_len);
   return !res;
 }
 
@@ -61,7 +59,7 @@ int64_t __string__less(void *lhs, void *rhs) {
   uint64_t lhs_len = *((int64_t *)lhs - 1);
   uint64_t rhs_len = *((int64_t *)rhs - 1);
   uint64_t len = lhs_len < rhs_len ? lhs_len : rhs_len;
-  int res = memcmp(lhs, rhs, len);
+  int64_t res = memcmp(lhs, rhs, len);
   if (res == 0)
     return lhs_len < rhs_len ? 1 : 0;
   return res < 0;
@@ -73,15 +71,15 @@ int64_t __string__less_equal(void *lhs, void *rhs) {
 
 void __string___ctor_(void *ptr) { *(uint64_t *)ptr = 0; }
 
-int getInt() {
+int64_t getInt() {
   char c = getchar();
-  int neg = 0;
+  int64_t neg = 0;
   while (c < '0' || c > '9') {
     if (c == '-')
       neg = 1;
     c = getchar();
   }
-  int num = c - '0';
+  int64_t num = c - '0';
   c = getchar();
   while (c >= '0' && c <= '9') {
     num = num * 10 + c - '0';
@@ -94,26 +92,26 @@ int getInt() {
 
 void print(void *ptr) {
   uint64_t len = *((int64_t *)ptr - 1);
-  printf("%.*s", (int)len, (char *)ptr);
+  printf("%.*s", (int64_t)len, (char *)ptr);
 }
 
-void _printInt(int num) {
+void _printInt(int64_t num) {
   if (num == 0)
     putchar('0');
   if (num < 0) {
     num = -num;
     putchar('-');
   }
-  int digits[10], len = 0;
+  int64_t digits[10], len = 0;
   while (num > 0) {
     digits[len++] = num % 10;
     num /= 10;
   }
-  for (int i = len - 1; i >= 0; --i)
+  for (int64_t i = len - 1; i >= 0; --i)
     putchar('0' + digits[i]);
 }
 
-void _printlnInt(int x) {
+void _printlnInt(int64_t x) {
   _printInt(x);
   putchar('\n');
 }
